+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Change Log
 
-## [Unreleased]
+## [0.3.0] 2022-08-05
 ### Added
+- `extraContext:editorSelectionStartVisible`
+- `extraContext:editorSelectionStartLineRelativeVisibleTop`
+- `extraContext:editorSelectionStartLineRelativeVisibleBottom`
 
 ## [0.2.0] 2022-04-03
 ### Added